requirements.txt
# requirements.txt
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-generativeai==0.3.2
pandas==2.1.3
python-dotenv==1.0.0
email-validator==2.1.0